% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_candidate.R
\name{find_top_candidate}
\alias{find_top_candidate}
\title{find_top_candidate}
\usage{
find_top_candidate(species_one_data, species_two_data, orthologs,
  the_quantiles, binomial_cuts)
}
\arguments{
\item{species_one_data}{list containing overall data in first position and tcontigs in second, for first species}

\item{species_two_data}{list containing overall data in first position and tcontigs in second, for second species}

\item{orthologs}{orthologs present in both, which have been cleaned}

\item{binomial_cuts}{the binomial cuts to test}

\item{quantiles}{the quantiles being tested}
}
\value{
returns a list containing the top candidates. First position is candidate overlap, second is candidates in species one, third is candidates in species 2
}
\description{
find_top_candidate
}
\examples{
find_top_cand(pine, spruce, cleaned_ortho, binomial_cuts)
}
\keyword{topcandidate}
