% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_cleaning.R
\name{ortholog_cleaning}
\alias{ortholog_cleaning}
\title{ortholog_cleaning}
\usage{
ortholog_cleaning(species_one_tcontigs, species_two_tcontigs, minimum)
}
\arguments{
\item{species_one_tcontigs}{contig data from species one}

\item{species_two_tcontigs}{contig data from species two}

\item{minimum}{the minimum number of SNP's needed for a gene}
}
\value{
returns the list of orthlogs that meet conditions and are in both species
}
\description{
Function to clean SNP data for two species entered
will get the unique names of each gene in species then find which ones have above a certain number of SNP's
Then returns the list representing the ortologs meeting this criteria that are in both species
}
\examples{
ortholog_cleaning(spruce_tcontig, pine_tcontig, 3)
}
\keyword{clean}
