% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_import.R
\name{ortholog_import}
\alias{ortholog_import}
\title{ortholog_import}
\usage{
ortholog_import(all_orthologs, one_to_one_orthologs)
}
\arguments{
\item{all_orthologs}{table formatted data containing all the orthologs between two species}

\item{one_to_one_orthologs}{table formatted data containine the one to one orthologs of these species}
}
\value{
returns a list of data Frames of the read in ortholog data
}
\description{
This function allows you to import all the necessary ortholog information between two species
Takes two table of ortholog sequences, one of which is all orthologs, and the other just the one to one
}
\examples{
ortholog_import("all_orthologs_pine_spruce.txt","one_to_one_orthologs.txt")
}
\keyword{orthologs}
