% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_data_import.R
\name{species_data_import}
\alias{species_data_import}
\title{experimental_data_import}
\usage{
species_data_import(overallInfo = NULL, get_only_annotated = TRUE,
  annotatedColumn = NULL, commentChar = "&")
}
\arguments{
\item{get_only_annotated}{will get only the SNP's annotated to a gene if true, defaults to TRUE}

\item{annotatedColumn}{the column which you want to ensure that each value is annotated to. As an example this could be the contig column if you only wanted SNPS annotated to a gene. Defaults to NULL}

\item{commentChar}{the character which indicates comments in the input data frame, defaults to &}

\item{overall_info}{required table formatted data containing all the experimental info about a species. Defaults to NULL}
}
\description{
Allows input of experimental info about a species
}
\examples{
species_data_import("all_pine_data_pvalues.txt", TRUE)


}
\keyword{species}
