% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_import.R
\name{species_data_import}
\alias{species_data_import}
\title{species_data_import}
\usage{
species_data_import(overallInfo, get_only_annotated = TRUE)
}
\arguments{
\item{get_only_annotated}{will get only the SNP's annotated to a gene if true, defaults to TRUE}

\item{overall_info}{table formatted data containing all the experimental info about a species}
}
\description{
Allows input of experimental info about a species
}
\examples{
species_data_import("all_pine_data_pvalues.txt", TRUE)
}
\keyword{species}
